requirements.txt
openai-whisper
torch
torchaudio
jiwer
fastapi
uvicorn
python-multipart
numpy